*** empty log message ***

+++ trunk/castor-2002/castor/src/examples/jdo/sybase.sql
@@ -9,7 +9,7 @@ go
 
 
 create table prod_group (
-  id    numeric       not null primary key,
+  id    int           not null primary key,
   name  varchar(200)  not null
 )
 go
@@ -20,7 +20,7 @@ create table prod (
   id        int           not null primary key,
   name      varchar(200)  not null,
   price     float         not null,
-  group_id  numeric       not null references prod_group( id )
+  group_id  int           not null references prod_group( id )
 )
 go
 grant all on prod to test
